Refactored namespaces
