+++ SharpDactyl.Cons/Program.cs
@@ -1,28 +1,28 @@
 ﻿using System;
-using System.Diagnostics;
 using System.Threading.Tasks;
-using SharpdactylLib.Models.Client;
+using Newtonsoft.Json;
+using SharpdactylLib.Models.Application.Server;
 
 namespace SharpDactyl.Cons
 {
     class Program
     {
-        static async Task Main(string[] args)
+        static async Task Main()
         {
-            SharpdactylLib.SharpDactyl test = new SharpdactylLib.SharpDactyl("https://panel.pcbyte.eu", "coM5TnrZNTBij1Ptxkc5FPK93JZD5yqUBZ8vPx1B0pZ5x1Bd");
+            SharpdactylLib.SharpDactyl test = new SharpdactylLib.SharpDactyl("[Domain]", "[Application API Key]");
             var result = await test.Application.GetServers();
 
             if(result.Count > 0){
                 foreach (ServerDatum serverDatum in result)
                 {
-                    Console.WriteLine($"{serverDatum.Attributes.Name} => {serverDatum.Attributes.Uuid}");
+                    string json = JsonConvert.SerializeObject(serverDatum, Formatting.Indented);
+                    Console.WriteLine($"{json}\n\n");
                 }
             }
             else
             {
                 Console.WriteLine($"No results");
             }
-            Console.ReadLine();
         }
     }
 }
