+++ Sharpdactyl/SharpDactyl.cs
@@ -6,9 +6,9 @@ using System.Threading.Tasks;
 using SharpdactylLib.Enums;
 using SharpdactylLib.Exceptions;
 using SharpdactylLib.Helpers;
-using SharpdactylLib.Models.Server;
-using SharpdactylLib.Models.Node;
-using SharpdactylLib.Models.User;
+using SharpdactylLib.Models.Application.Server;
+using SharpdactylLib.Models.Application.Node;
+using SharpdactylLib.Models.Application.User;
 
 namespace SharpdactylLib
 {
