+++ Sharpdactyl/Sharpdactyl.csproj
@@ -28,4 +28,8 @@
     <PackageReference Include="Flurl.Http" Version="3.0.0-pre4" />
     <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
   </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Models\Client\" />
+  </ItemGroup>
 </Project>
